Feature/product testcode

+++ module-domain/src/main/java/com/kernel360/auth/entity/Auth.java
@@ -17,10 +17,10 @@ public class Auth extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "auth_id_gen")
     @SequenceGenerator(name = "auth_id_gen", sequenceName = "auth_auth_no_seq")
     @Column(name = "auth_no", nullable = false)
-    private Integer authNo;
+    private Long authNo;
 
     @Column(name = "member_no")
-    private Integer memberNo;
+    private Long memberNo;
 
     @Column(name = "jwt_token")
     private String jwtToken;
@@ -32,7 +32,7 @@ public class Auth extends BaseEntity {
     /**
      * all create
      **/
-    public static Auth of(Integer authNo, Integer memberNo, String jwtToken, String snsToken) {
+    public static Auth of(Long authNo, Long memberNo, String jwtToken, String snsToken) {
 
         return new Auth(authNo, memberNo, jwtToken, snsToken);
     }
@@ -40,7 +40,7 @@ public class Auth extends BaseEntity {
     /**
      * all binding
      **/
-    private Auth(Integer authNo, Integer memberNo, String jwtToken, String snsToken) {
+    private Auth(Long authNo, Long memberNo, String jwtToken, String snsToken) {
         this.authNo = authNo;
         this.memberNo = memberNo;
         this.jwtToken = jwtToken;
@@ -48,7 +48,7 @@ public class Auth extends BaseEntity {
     }
 
 
-    public static Auth jwt(Integer authNo, Integer memberNo, String jwtToken) {
+    public static Auth jwt(Long authNo, Long memberNo, String jwtToken) {
 
         return new Auth(authNo, memberNo, jwtToken);
     }
@@ -56,7 +56,7 @@ public class Auth extends BaseEntity {
     /**
      * JWT
      **/
-    private Auth(Integer authNo, Integer memberNo, String jwtToken) {
+    private Auth(Long authNo, Long memberNo, String jwtToken) {
         this.authNo = authNo;
         this.memberNo = memberNo;
         this.jwtToken = jwtToken;
